refactor(ErrorCode): 각 도메인에서 ErrorCode를 관리하도록 수정
